done with automation

+++ go_fetch.py
@@ -22,13 +22,14 @@ if not (file_exists := os.path.exists(file_location)):
         case 200:
             if not os.path.exists(file_location[:-9]):
                 os.makedirs(file_location[:-9])
-            with open(file_location, 'w') as f:
-                f.write(r.text)
             print(f'Retrieval succeeded.')
-            cmd = f'copy \"{os.getcwd()}\\{year}\\_new\\rename_me.py\" \"{os.getcwd()}\\{year}\\day{day}\\day{day}.py\"'
-            with open(file_location[:-9] + 'test.txt', 'w') as f:
-                f.write("GO FIND THE TEST INPUT LOSER")
-            os.system(cmd)
+            if not os.path.exists(file_location[:-9] + f'day{day}.py'):
+                cmd = f'copy \"{os.getcwd()}\\{year}\\_new\\rename_me.py\" \"{os.getcwd()}\\{year}\\day{day}\\day{day}.py\"'
+                os.system(cmd)
+            if not os.path.exists(file_location[:-9] + f'test.txt'):
+                with open(file_location[:-9] + 'test.txt', 'w') as f:
+                    f.write("GO FIND THE TEST INPUT LOSER")
+
         case _ as status_code:
             print(f'Retrieval failed: Status Code:{status_code}')
 else:
